Move dumb components to /components
