+++ src/renderer/views/App.tsx
@@ -7,7 +7,7 @@ import BrushTool from "../tools/BrushTool"
 import WatercolorTool from "../tools/WatercolorTool"
 import DrawArea from "./DrawArea"
 import LayerList from "./LayerList"
-import ColorPicker from "./ColorPicker"
+import ColorPicker from "./components/ColorPicker"
 import Palette from "./Palette"
 import Navigator from "./Navigator"
 import RGBRangeSliders from "./components/RGBRangeSliders"
